requirements.txt
fastapi
uvicorn
pydantic
httpx
google-adk